+++ internal/component/azure/blobstorage/metadata_test.go
@@ -19,12 +19,9 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
 	"github.com/stretchr/testify/assert"
-
-	"github.com/dapr/kit/logger"
 )
 
 func TestParseMetadata(t *testing.T) {
-	logger := logger.NewLogger("test")
 	var m map[string]string
 
 	t.Run("parse all metadata", func(t *testing.T) {
@@ -78,17 +75,5 @@ func TestParseMetadata(t *testing.T) {
 		}
 		_, err := parseMetadata(m)
 		assert.Error(t, err)
-	})
-
-	t.Run("sanitize metadata if necessary", func(t *testing.T) {
-		m = map[string]string{
-			"somecustomfield": "some-custom-value",
-			"specialfield":    "special:valueÜ",
-			"not-allowed:":    "not-allowed",
-		}
-		meta := SanitizeMetadata(logger, m)
-		assert.Equal(t, meta["somecustomfield"], "some-custom-value")
-		assert.Equal(t, meta["specialfield"], "special:value")
-		assert.Equal(t, meta["notallowed"], "not-allowed")
 	})
 }
